% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{add_shortnames}
\alias{add_shortnames}
\title{Add parameter aliases (generalized short names) and other fields}
\usage{
add_shortnames(x)
}
\arguments{
\item{x}{A data.frame of class \code{ec_tidy} or \code{ems_tidy}}
}
\value{
Same as input but with additional fields: "VMV_CODE", "PARAM_SHORT_NAME",
"VARIABLE_GROUP", "PARAM_GROUP", "CONSTIT_ABBREV"
}
\description{
Add parameter aliases (generalized short names) and other fields
}
