% Generated by roxygen2 (4.0.2): do not edit by hand
\name{as_if_wq_thresholds}
\alias{as_if_wq_thresholds}
\title{Test if equivalent to wq_thresholds object}
\usage{
as_if_wq_thresholds(x)
}
\arguments{
\item{x}{object to test}
}
\value{
flag if successful otherwise an informative error message
}
\description{
Tests if object is equvialent to an object of
 class wq_thresholds
}

