% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-daily-cv.R
\name{calc_daily_cv}
\alias{calc_daily_cv}
\title{Calculates Daily Coefficient of Variation}
\usage{
calc_daily_cv(x, messages = getOption("wqbc.messages", default = TRUE),
  parallel = getOption("wqbc.parallel", default = FALSE))
}
\arguments{
\item{x}{data.frame with columns Variable, Value, Units and Date.}

\item{messages}{flag indicating whether to print messages}

\item{parallel}{flag indicating whether to calculate cvs
using the parallel backend provided by foreach}
}
\description{
Calculates the daily coefficient of variation for values for
the same variable collected on the same date. Useful for identifying dates
in which there was substantial variation between replicates.
Values will be convrrted
}
\examples{
data <- data.frame(Variable = "A Variable", Date = Sys.Date(),
                   Value = c(20, 1, 2), Units = c("mg/L", "mg/L", "mg/L"))

data <- rbind(data, data.frame(Variable = "Another Variable", Date = Sys.Date(),
                   Value = c(1, 10^3), Units = c("mg/L", "ug/L")))

calc_daily_cv(data, messages = TRUE)
}

