% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-rcvs.R
\name{calc_rcvs}
\alias{calc_rcvs}
\title{Calculates Replicates Coefficients of Variation}
\usage{
calc_rcvs(x, messages = getOption("wqbc.messages", default = TRUE),
  parallel = getOption("wqbc.parallel", default = FALSE))
}
\arguments{
\item{x}{data.frame with columns Variable, Value, Units and Date.}

\item{messages}{flag indicating whether to print messages}

\item{parallel}{flag indicating whether to calculate cvs
using the parallel backend provided by foreach}
}
\description{
Calculates the coefficient of variation for values for
the same variable collected on the same date. Useful for identifying dates
in which there was substantial variation between replicates.
If possible replicates with inconsistent units are converted to the same
units - otherwse
}
\examples{
data(dummy)
calc_rcvs(dummy, messages = TRUE)
}

