% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqi.R
\name{calc_wqi}
\alias{calc_wqi}
\title{Calculate Water Quality Index (WQI)}
\usage{
calc_wqi(x, by = NULL, ci = "row", cesi_code = FALSE,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{A data.frame to calculate the WQI for.}

\item{by}{An optional character vector of the columns in x to calculate the WQI by.}

\item{ci}{A string indicating whether to calculate 95% bootstrap confidence
intervals by "row" or "column" or "none". Currently the confidence intervals are
considered unreliable and should not be used.}

\item{cesi_code}{A temporary flag indicating whether to use wqbc or cesi code to calculate the
confidence intervals. This argument should be removed prior to release of the package.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Calculates a water quality index for a series of variables, values with upper limits
and optionally lower limits using the method detailed in the
CCME Water Quality Index 1.0 User's Manual.
}
\details{
The upper limits can be generated using the \code{\link{calc_limits}} function
or can be provided by the user. In fact if x lacks both upper and lower limits
then the \code{calc_limits} function is automatically called prior to
calculating the WQI. If values are zero and detection limits are provided
then the values are set to be the detection limits. This is important when
the variable has lower limits because otherwise the excursion will be infinity
and it will not be possible to calculate the WQI. In this case \code{calc_wqi}
throws an informative error. Finally it is important to note that in order
for the WQI to be calculated the data set must include four variables each
with non-missing values on at least four separate days.
}
\examples{
data(ccme)
calc_wqi(ccme, messages = TRUE)
calc_wqi(ccme, by = "Date", messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{wqbc}}
}

