% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{categorize_wqi}
\alias{categorize_wqi}
\title{Categorize WQI Values}
\usage{
categorize_wqi(x)
}
\arguments{
\item{x}{numeric vector of WQI values}
}
\value{
factor of WQI categories
}
\description{
Categorize WQI Values
}
\examples{
categorize_wqi(seq(0, 100, by = 5))
}

