% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccme_raw}
\alias{ccme_raw}
\title{Raw CCME Water Quality Index User’s Manual Example Data}
\format{A data frame with 12 rows and 11 variables:
\describe{
  \item{Date}{date of reading}
  \item{DO}{dissolved oxygen value}
  \item{pH}{pH value}
  \item{TP}{total phosphorus value}
  \item{TN}{total nitrogen value}
  \item{FC}{faecal coliform value}
  \item{As}{arsenic value}
  \item{Pb}{lead value}
  \item{Hg}{mercury value}
  \item{x2.4.D}{2,4-D value}
  \item{Lindane}{linedane value}
}}
\usage{
ccme_raw
}
\description{
Example raw dataset from the CCME Water Quality Index 1.0 User’s Manual
}
\keyword{datasets}

