% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean-wqdata.R
\name{clean_wqdata}
\alias{clean_wqdata}
\title{Clean Water Quality Data}
\usage{
clean_wqdata(x, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{data.frame with columns Date, Variable, Value and Units}

\item{messages}{flag indicating whether to print messages}
}
\description{
Cleans water quality data.
}
\examples{
calc_rcvs(dummy, messages = TRUE)
}

