% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\title{Water Quality Parameter codes and units}
\format{A data frame with 3 variables:
\describe{
  \item{Code}{unique short-hand code}
  \item{Variable}{unique name of water quality parameter}
  \item{Units}{units for parameter}
  \item{Average}{R function to calculate "average" value for multiple samples in a period}
}}
\usage{
codes
}
\description{
Water Quality Parameter codes and units
}
\details{
Water Quality Parameter codes and units
}
\keyword{datasets}

