% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/convert-ems-data.R
\name{convert_ems_data}
\alias{convert_ems_data}
\title{Convert EMS Data}
\usage{
convert_ems_data(x, date = "date", code = "code", value = "value",
  units = "units", messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The data.frame to convert.}

\item{date}{A string of Date column name.}

\item{code}{A string of Code column name.}

\item{value}{A string of Value column name.}

\item{units}{A string of Units column name.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Converts EMS data.
}
\examples{
data <- data.frame(date = as.Date("2000-01-01"), code = c("SE-T", "KRYP", "0004"),
value = 1, units = c("MG/L","MG/L","PH UNITS"))
convert_ems_data(data)
}

