% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/units.R
\name{convert_units}
\alias{convert_units}
\title{Convert Units}
\usage{
convert_units(x, from, to, messages = TRUE)
}
\arguments{
\item{x}{numeric vector of values to convert}

\item{from}{character vector of original units}

\item{to}{character vector new units}

\item{messages}{flag indicating whether to print messages}
}
\value{
numeric vector of values in new units
}
\description{
Converts units
}
\examples{
convert_units(1:10, from = "mg/L", to = "ug/L")
}

