% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/units.R
\name{convert_values}
\alias{convert_values}
\title{Convert Values}
\usage{
convert_values(x, from, to)
}
\arguments{
\item{x}{numeric vector of values to convert}

\item{from}{character vector of original units}

\item{to}{character vector new units}
}
\value{
numeric vector of values in new units
}
\description{
Converts values from original units to new units
}
\examples{
convert_values(1:2, from = "mg/L", to = "mg/L")
convert_values(1:2, from = "mg/L", to = "ug/L")
convert_values(1:2, from = "mg/L", to = "midichlorians")
convert_values(1:2, from = "midichlorians", to = "midichlorians")
}

