% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy}
\alias{dummy}
\title{Dummy Water Quality Data}
\format{A data frame with 5 columns:
\describe{
  \item{ID}{row identifier}
  \item{Variable}{parameter variable name}
  \item{Date}{date of reading}
  \item{Value}{value of reading}
  \item{Units}{units of reading}
}}
\usage{
dummy
}
\description{
A dummy data set to illustrate various data cleaning functions.
}
\examples{
demo(dummy, ask = FALSE)
}
\keyword{datasets}

