% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ems_codes}
\alias{ems_codes}
\title{Water Quality Parameter EMS Names and Codes for British Columbia}
\format{
A tibble with 4 variables:
\describe{
\item{Variable}{The name of the variable.}
\item{Code}{The EMS code.}
}
}
\usage{
ems_codes
}
\description{
The standard variables and codes stored in the EMS database.
}
\keyword{datasets}
