% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{error}
\alias{error}
\title{Error}
\usage{
error(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}
}
\description{
Throws an error without the call as part of the error message.
}
\seealso{
base::stop
}

