% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codes.R
\name{expand_ems_codes}
\alias{expand_ems_codes}
\title{Expand EMS Codes}
\usage{
expand_ems_codes(x)
}
\arguments{
\item{x}{A character vector of codes to expand}
}
\description{
Expands EMS codes by adding EMS_ to start if absent
and replacing all '-' with '_'. This function is provided
because wqbc stored EMS codes in expanded form.
}
\examples{
expand_ems_codes(c("0014", "KR-P", "0_15", "EMS_ZN_T"))
}
\seealso{
\code{\link{compress_ems_codes}}
}

