% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fraser}
\alias{fraser}
\title{Fraser River Basin Long-term Water Quality Monitoring 1979-Present}
\format{A data frame with 8 columns:
\describe{
\item{SiteID}{The unique water quality station number.}
\item{Date}{The date of the reading.}
\item{Variable}{The name of the variable.}
\item{Value}{The value of the reading.}
\item{Units}{The units of the value.}
\item{Site}{The full name of the station.}
\item{Lat}{The latitude of the station in decimal degrees.}
\item{Long}{The longitude of the station in decimal degrees.}
}}
\source{
\url{http://open.canada.ca/data/en/dataset/9ec91c92-22f8-4520-8b2c-0f1cce663e18}
}
\usage{
fraser
}
\description{
Surface freshwater quality monitoring in the Fraser River Basin
is carried out under the Canada-British Columbia Water Quality
Monitoring Agreement. Monitoring is conducted to assess water
quality status and long-term trends, detect emerging issues,
establish water quality guidelines and track the effectiveness
of remedial measures and regulatory decisions.
}
\examples{
\dontrun{
demo(fraser)
}
}
\keyword{datasets}
