% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/average.R
\name{geomean1}
\alias{geomean1}
\title{Geometric Mean Plus-Minus 1}
\usage{
geomean1(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector of non-negative numbers.}

\item{na.rm}{A flag indicating whether to remove missing values.}
}
\description{
Calculates the geometric mean by adding 1 before logging
and subtracting 1 before exponentiating so that provides results
even with zero counts. Not used by any wqbc functions but provided
as may be helpful if averaging bacterial counts.
}
\examples{
mean(0:9)
geomean1(0:9)
}

