% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_category_colors}
\alias{get_category_colors}
\title{Get Category Colors}
\usage{
get_category_colors()
}
\description{
Returns a named vector of the category colors to use when
plotting water quality index values.
}
\examples{
get_category_colors()
}
\seealso{
\code{\link{calc_wqis}}
}

