% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_category_colours}
\alias{get_category_colours}
\title{Get Category Colours}
\usage{
get_category_colours()
}
\value{
named character vector of colours for water quality index categories
}
\description{
Returns a named vector of the default category colours to use when
plotting water quality index values.
}
\examples{
get_category_colours()
}
\seealso{
\code{\link{calc_wqis}}
}

