% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_codes_variables}
\alias{get_codes_variables}
\title{Get Water Quality Code-Variable Lookup}
\usage{
get_codes_variables()
}
\description{
Returns a data.frame of the water quality code and variable
look up table currently defined in the wqbc package.
}
\examples{
get_codes_variables()
}

