% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{get_excursion}
\alias{get_excursion}
\title{Get Excursion}
\usage{
get_excursion(value, lower = NA_real_, upper = NA_real_)
}
\arguments{
\item{value}{numeric vector of values to calculate excursion}

\item{lower}{numeric vector of lower limits}

\item{upper}{numeric vector of upper limits}
}
\value{
numeric vector of excursions
}
\description{
Get Excursion
}
\examples{
library(dplyr)
data(ccme)
ccme$Excursion <- get_excursion(ccme$Value, ccme$LowerLimit, ccme$UpperLimit)
dplyr::filter(ccme, Excursion != 0)
}

