% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-guidelines.R
\name{get_guidelines}
\alias{get_guidelines}
\title{Get Water Quality Guidelines}
\usage{
get_guidelines(x, use = "Freshwater Life", jurisdiction = "BC")
}
\arguments{
\item{x}{data.frame with column Variables or Code which}

\item{use}{string of required use}

\item{jurisdiction}{string of regulatory body}
}
\description{
Gets water quality thresholds/objectives for British Columbia (BC) or
Canada (CA) for a range
}
\details{
If the argument \code{x} is missing the function returns all
the guidelines for all the variables currently defined in the
wqbc package for the specified use and jurisdiction.
}
\examples{
get_guidelines()
get_guidelines(use = "Drinking")
}

