% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_uses}
\alias{get_uses}
\title{Get Water Quality Uses}
\usage{
get_uses()
}
\description{
Returns a character vector of the uses for which
limits are currently defined in the wqbc package.
}
\examples{
get_uses()
}

