% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-wqdata.R
\name{get_wqdata}
\alias{get_wqdata}
\title{Get WQ Data}
\usage{
get_wqdata(emsid = NULL, param_code = NULL, from_date = NULL,
  to_date = NULL, force = FALSE, ask = TRUE)
}
\arguments{
\item{emsid}{A character vector of the ems id(s) of interest}

\item{param_code}{a character vector of parameter codes}

\item{from_date}{A date string in a standard unambiguous format (e.g., "YYYY/MM/DD")}

\item{to_date}{A date string in a standard unambiguous format (e.g., "YYYY/MM/DD")}

\item{force}{Force downloading the dataset, even if it's not out of date (default \code{FALSE})}

\item{ask}{should the function ask for your permission to cache data on your computer?
Default \code{TRUE}}
}
\value{
A data.frame of water quality data suitable for use by wqbc functions.
}
\description{
A wrapper on rems functions to get EMS data from
}
\seealso{
\code{\link[rems]{download_historic_data}} and \code{\link[rems]{read_historic_data}}
}

