% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify Outliers In Water Quality Data}
\usage{
identify_outliers(data, by = NULL, threshold = 100, time_series = FALSE,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{by}{A character vector of the columns in x to perform the outlier detection by.}

\item{threshold}{A number indicating the maximum permitted coefficient}

\item{time_series}{A flag indicating whether to identify outliers using a time series model.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Identifies outliers in water quality data.
}
\details{
The method is motivated by Hampels Mean Absolute Deviation approach for time series. It is a
combination of two techniques.
First it identifies outliers in exceedance of the threshold.
Then if time_series = TRUE it
    a. assigns zero weight to outliers.
    b. fits a reasonably flexible time series model to identify
       trend in the data using a robust regression approach by iteratively reweighting
    b. assign zero weight to observations with very large residuals from the robust fit
    c. repeat this procedure until all outliers have been identified.
}
\examples{
data <- standardize_wqdata(wqbc::dummy)
identify_outliers(data, by = "Variable", threshold = 10, messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}

