% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{is_within_limits}
\alias{is_within_limits}
\title{Is Within Limits}
\usage{
is_within_limits(value, lower = NA_real_, upper = NA_real_)
}
\arguments{
\item{value}{numeric vector of values to test}

\item{lower}{numeric vector of lower limits}

\item{upper}{numeric vector of upper limits}
}
\value{
logical vector indicating whether within limits
}
\description{
Is Within Limits
}
\examples{
library(dplyr)
data(ccme)
ccme$Within <- is_within_limits(ccme$Value, ccme$LowerLimit, ccme$UpperLimit)
filter(ccme, !Within)
}

