% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.R
\name{is_wq_thresholds}
\alias{is_wq_thresholds}
\title{Test wq_thresholds}
\usage{
is_wq_thresholds(x)
}
\arguments{
\item{x}{object to test}
}
\value{
flag indicating result
}
\description{
Test if object is of class wq_thresholds
}

