% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limits}
\alias{limits}
\title{Water Quality Limits for British Columbia and Canada}
\format{A data frame with 13 variables:
\describe{
  \item{Variable}{name of water quality parameter}
  \item{Period}{Period for which limit applies. When multiple values occur
  within the same period the average function in the codes data.frame is
  used to calculate the average value. Note in order for the month limits to
  apply at least five values must have been recorded on five separate days
  from at least three calendar weeks within the calendar month}
  \item{Condition}{R logical expression to test required condition}
  \item{LowerLimit}{R expression defining lower limit}
  \item{UpperLimit}{R expression defining upper limit}
  \item{Units}{units for limit}
}}
\usage{
limits
}
\description{
Water Quality Limits for British Columbia and Canada
}
\keyword{datasets}

