% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_codes}
\alias{lookup_codes}
\title{Get Codes}
\usage{
lookup_codes(variables = NULL, messages = getOption("wqbc.messages", default
  = TRUE))
}
\arguments{
\item{variables}{An optional character vector of variables to get codes for.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Gets the recognised water quality codes.
}
\examples{
lookup_codes()
lookup_codes(c(lookup_variables()[1:3], "Kryptonite"))
}

