% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_limits}
\alias{lookup_limits}
\title{Lookup Limits}
\usage{
lookup_limits(ph = NA_real_, hardness = NA_real_, chloride = NA_real_,
  methyl_mercury = NA_real_, term = "long")
}
\arguments{
\item{ph}{A number indicating the pH in pH units at the site of interest.}

\item{hardness}{A number indicating the total hardness (CaCO3) in mg/L at the site of interest.}

\item{chloride}{A number indicating the total chloride concentration in mg/L at the site of interest.}

\item{methyl_mercury}{A number indicating the total concentration of methyl mercury in ug/L at the site of interest.}

\item{term}{A string indicating whether to lookup the "long" or "short"-term limits.}
}
\description{
Looks up the long or short-term water quality limits for BC. If the limits depend on
on the pH, total hardness (CaCO3), total chloride or the concentration of methyl mercury
and site specific values are not provided then the dependent limits are returned
as missing values.
}
\examples{
lookup_limits(ph = 8, hardness = 100, chloride = 50, methyl_mercury = 2)
lookup_limits(ph = 8, hardness = 100, chloride = 50, methyl_mercury = 2, term = "short")
}

