% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_units}
\alias{lookup_units}
\title{Lookup Units}
\usage{
lookup_units()
}
\description{
Returns a character vector of the recognised units.
}
\examples{
lookup_units()
}
\seealso{
\code{\link{lookup_limits}}
}

