% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map}
\alias{map}
\title{Borders of British Columbia}
\format{A data frame with 4,920 rows and 3 variables:
\describe{
  \item{Longitude}{longitude}
  \item{Latitude}{latitude}
  \item{Group}{grouping variable for plotting to ensure discrete polygons}
}}
\usage{
map
}
\description{
Borders of British Columbia used by \code{\link{plot_map}} function.
}
\keyword{datasets}

