% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outlier_id}
\alias{outlier_id}
\title{Identify Outliers In Water Quality Data}
\usage{
outlier_id(x, by = NULL, mad_threshold = 10, ts_threshold = 6,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The data.frame to analyse.}

\item{by}{A character vector of the columns in x to perform the outlier detection by.}

\item{mad_threshold}{A number indicating the maximum permitted coefficient}

\item{ts_threshold}{A number indicating the maximum permitted coefficient}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Identifies outliers in water quality data.
}
\details{
The method is motivated by Hampels Mean Absolute Deviation approach for time series. It is a 
combination of two tecniques:
  1. remove large outliers when compared to the MAD of the whole time series
  2. 
    a. fit a reasonably flexible time series model to identify 
       trend in the data using a robust regression approach by iteratively reweighting
    b. assign zero weight to observations with very large residuals from the robust fit
    c. repeat this procedure until all outliers have been identified.
}
\examples{
x <- standardize_wqdata(wqbc::dummy)
outlier_id(x, by = "Variable", messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}

