% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detection.R
\name{outlier_removal}
\alias{outlier_removal}
\title{Remove Outliers From Water Quality Data}
\usage{
outlier_removal(x, by = NULL, w_threshold = 6, theshold = 10,
  max_cv = 1.29, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The data.frame to clean.}

\item{by}{A character vector of the columns in x to perform the cleaning by.}

\item{max_cv}{A number indicating the maximum permitted coefficient
of variation.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Removes outliers from water quality data.
}
\details{
The method is based on Hampels Mean Absolute Deviation approach for time series. It is a 
combination of two tecniques:
  1. remove large outliers when compared to the MAD of the whole time series
  2. 
    a. fit a reasonably flexible time series model to identify 
       trend in the data using a robust regression approach by iteratively reweighting
    b. delete observations with very large residuals from the robust fit
    c. repeat this procedure until all outliers have been removed.
}
\examples{
outlier_removal(wqbc::dummy, by = "Variable", messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}

