% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map_wqis}
\alias{plot_map_wqis}
\title{Plot Map of Water Quality Indices}
\usage{
plot_map_wqis(data, x = "Long", y = "Lat", size = 3, shape = 21,
  theme = theme_map(), drop = TRUE, input_proj = NULL,
  palette = getOption("wqbc.category_colours", get_category_colours()))
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{y}{string of column in data to plot on y axis}

\item{size}{number of point size or string of column in data
to plot size of points}

\item{shape}{integer of point shape (permitted values are 21 to 25)
or string of column in data to plot shape of points}

\item{theme}{ggplot theme}

\item{drop}{flag indicating whether to drop duplicated rows to
avoid overplotting}

\item{input_proj}{a valid proj4string. Defaults to longlat/NAD83
(\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"})}

\item{palette}{named character vector of palette for fill of points}
}
\description{
Creates ggplot2 object with map polygon,
coord_fixed and theme_minimal and fill based
on WQI categories.
}
\examples{
\dontrun{
 demo(fraser)
}
}

