% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries_by}
\alias{plot_timeseries_by}
\title{Plot Data By}
\usage{
plot_timeseries_by(data, by = c("Site", "Variable"), x = "Date",
  y = "Value", xlab = x, ylab = y, color = "black", y0 = TRUE)
}
\arguments{
\item{data}{A data frame of the data to plot.}

\item{by}{A character vector of the columns to plot the time series by.}

\item{x}{A string of the column to plot on the x-axis.}

\item{y}{A string of the column to plot on the y-axis.}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the y-axis label.}

\item{color}{A string specifying the color for the points or if not a color the name of the column to color the points by.}

\item{y0}{A flag indicating whether to expand the y-axis limits to include 0.}
}
\value{
A list of ggplot objects
}
\description{
Plot Data By
}

