% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wqis}
\alias{plot_wqis}
\title{Plot Water Quality Indices}
\usage{
plot_wqis(data, x = "Tests", size = 3, shape = 21, theme = theme_wqis(),
  palette = getOption("wqbc.category_colours", get_category_colours()))
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{size}{number of point size or string of column in data
to plot size of points}

\item{shape}{integer of point shape (permitted values are 21 to 25)
or string of column in data to plot shape of points}

\item{theme}{ggplot theme}

\item{palette}{named character vector of palette for fill of points}
}
\value{
ggplot2 object
}
\description{
Creates ggplot2 object with
y-limits expanded to include 0 and 100 and defined fill scale
for the WQI values corresponding by default to get_category_colours()
}
\examples{
library(ggplot2)

data(ccme)
plot_wqis(calc_wqis(ccme))
}
\seealso{
\code{\link{get_category_colours}}
}

