% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_limits}
\alias{site_limits}
\title{Example Site Specific Limits}
\format{
A tibble
\describe{
\item{Station_Name}{The name of the station as a character.}
\item{Variable}{The variable name as character.}
\item{Term}{The term ('Short' or 'Long') as a character.}
\item{Condition}{The condition as a character.}
\item{UpperLimit}{The upper limit as a character.}
\item{Units}{The units as a character.}
\item{Code}{The EMS codes as a character.}
\item{EMS_ID}{The station EMS ID as a character.}
\item{Use}{The use name as character.}
}
}
\usage{
site_limits
}
\description{
The standard variables, VMV codes and Variable Codes provided by Environment Canada.
}
\keyword{datasets}
