% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardise-values-units.R
\name{standardise_values_units}
\alias{standardise_values_units}
\title{Standardises Values and Units}
\usage{
standardise_values_units(x, messages = getOption("wqbc.messages", default =
  TRUE), parallel = getOption("wqbc.parallel", default = FALSE))
}
\arguments{
\item{x}{data.frame with columns Variable, Value and Units}

\item{messages}{flag indicating whether to print messages}

\item{parallel}{flag indicating whether to calculate limits by the by argument using the parallel backend provided by foreach}
}
\description{
Where possible standardises values and units
}
\examples{
data(dummy)
standardise_values_units(dummy, messages = TRUE)
}

