% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{standardize_wqdata}
\alias{standardize_wqdata}
\title{Standardize Water Quality Data}
\usage{
standardize_wqdata(x, strict = TRUE, messages = getOption("wqbc.messages",
  default = TRUE))
}
\arguments{
\item{x}{The data.frame to standardize.}

\item{strict}{A flag indicating whether to require all words
in a variable name to be present or only the first word.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Standardizes a water quality data set so that only recognised
variables and units remain and values have consistent units.
Negative or missing values are removed. When strict = FALSE
ambiguous variables such as "Iron Dissolved"
and "Iron Total" are dropped.
}
\examples{
standardize_wqdata(wqbc::dummy, messages = TRUE)
}

