% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations}
\alias{stations}
\title{Water Quality Stations for British Columbia}
\format{A tibble with 4 variables:
\describe{
\item{EMS_ID}{The station ID (chr).}
\item{Station_Name}{The EMS name of the station (chr).}
\item{Latitude}{The station latitude in decimal degrees (dbl).}
\item{Longitude}{The station longitude in decimal degrees (dbl).}
}}
\usage{
stations
}
\description{
The water quality stations for British Columbia with their coordinates.
}
\keyword{datasets}
