% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_colnames}
\alias{substitute_colnames}
\title{Substitute Column Names}
\usage{
substitute_colnames(x, messages = TRUE)
}
\arguments{
\item{x}{character vector of column names to substitute}

\item{messages}{flag indicating whether to print messages}
}
\value{
character vector of substituted or original names
}
\description{
Where possible substitute column names
with recognised column names
}
\examples{
substitute_colnames(c("dates", "variables", "latitude", "longitude"))
}

