% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_names}
\alias{substitute_names}
\title{Substitute Names}
\usage{
substitute_names(x, messages = TRUE)
}
\arguments{
\item{x}{data.frame}

\item{messages}{flag indicating whether to print messages}
}
\value{
data.frame
}
\description{
Where possible substitute  names
with recognised names
}
\examples{
x <- data.frame(variable = c("total phosphorus", "arsenic"),
               value = c(1, 2),
               units = c("MG / L", "Ug/l "))
substitute_names(x)
}

