% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_variables}
\alias{substitute_variables}
\title{Substitute Variable Names}
\usage{
substitute_variables(x, strict = TRUE, messages = getOption("wqbc.messages",
  default = TRUE))
}
\arguments{
\item{x}{The character vector of variable names to substitute.}

\item{strict}{A flag indicating whether to require all words
in a variable name to be present in or only
the first word one. Note when strict = FALSE
ambiguous variables such as "Iron Dissolved"
and "Iron Total" are dropped.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Where possible substitute variable names
with recognised variable names
}
\examples{
substitute_variables(c("ALUMINIUM SOMETHING", "FLUORIDE DISSOLVED",
"FLUORIDE", "NITROGEN DISSOLVED NITRATE", "PHOSPHORUS - TOTAL",
"KRYPTONITE", "OXYGEN", "OXYGEN DISSOLVED"))
}

