% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-trends.R
\name{summarise_for_trends}
\alias{summarise_for_trends}
\title{Summarise data by year and month}
\usage{
summarise_for_trends(data, breaks = NULL, FUN = "median",
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{breaks}{A numeric vector used to create groups of consecutive months, if NULL the full
year is used.}

\item{FUN}{The function to use for yearly summaries, e.g. median, mean, or max.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A tibble data.frame with rows for each Station Variable, year and month grouping.
}
\description{
Compute annual summaries of water quality observations.
}
\details{
The data must contain the columns Station, Date, Variable, Value, and Units.



This routine computes a prewhitened nonlinear trend on a vector of data, using either
Zhang's (described in Wang and Swail, 2001) or Yue Pilon's (describe in Yue Pilon, 2002)
method of prewhitening and Sen's slope, and use a Kendall test for significance.
}
\examples{
 data <- wqbc::yuepilon
 trend <- summarise_for_trends(data, breaks = 6)
 trend <- test_trends(data, breaks = 6, messages = TRUE, method = "sen")
}

