% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-wqdata.R
\name{summarise_wqdata}
\alias{summarise_wqdata}
\title{Summarise Water Quality Data}
\usage{
summarise_wqdata(x, by = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{The data.frame to summarise.}

\item{by}{A character vector of the columns in x to summarise by.}

\item{na.rm}{A flag specifying whether to exclude missing values when summarising.}
}
\value{
A tibble of the summary statistics.
}
\description{
Summarises water quality data in the Values (numeric) column
by Variable (character or factor) column.
}
\details{
Provides basic summary statistics (n, min, max, mean, median, se, sd).
SD; Options for quantiles, percentiles, upper/lower confidence intervals). View summary statistics in a tab of the results panel (added function to wqbc)
}
\examples{
data.frame(Variable = "var", Value = 1:5, stringsAsFactors = FALSE)
}
