% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{tidy_ems_data}
\alias{tidy_ems_data}
\title{Tidy EMS Data}
\usage{
tidy_ems_data(x, cols = character(0), mdl_action = "zero")
}
\arguments{
\item{x}{The rems data to tidy.}

\item{cols}{additional columns from the EMS data to retain.}

\item{mdl_action}{What to do with results that are below the detection limit.
Can be set to \code{zero} (the default), set at the detection limit (\code{mdl}),
or set to half the detection limit (\code{half}).}
}
\value{
A tibble of the tidied rems data.
}
\description{
Tidies water quality data downloaded from EMS database using the bcgov/rems package.
It retains and renames required columns and sets the timezone to PST.
}
\details{
It sets values that are flagged as being less than the detection limit to zero.
It does not alter values that are flagged as being greater than the detection limit -
that is left up to the user.
}
