% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-rems-data.R
\name{tidy_rems_data}
\alias{tidy_rems_data}
\title{Tidy Rems Data}
\usage{
tidy_rems_data(x)
}
\arguments{
\item{x}{The rems data to tidy.}
}
\value{
A tibble of the tidied rems data.
}
\description{
Tidies data from rems by retaining and renaming required columns and setting the
timezone to PST.
}

