% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vmv_codes}
\alias{vmv_codes}
\title{Water Quality Parameter VMV and Variable Codes for Canada}
\format{A tibble
\describe{
\item{Variable}{The name of the variable.}
\item{VMV_Code}{The VMV code.}
\item{EC_Code}{The Variable Code in the Environment Canada data.}
}}
\usage{
vmv_codes
}
\description{
The standard variables, VMV codes and Variable Codes provided by Environment Canada.
}
\details{
There can be more than one VMV Code for a variable!
}
\keyword{datasets}
