% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_add_codes}
\alias{wq_add_codes}
\title{Adds Code Column}
\usage{
wq_add_codes(x)
}
\arguments{
\item{x}{data.frame with Variable column}
}
\value{
data.frame with Variable and Code columns
}
\description{
Adds a Code column to the data.frame x based on the Variable column.
Unknown variable names are dropped.
}

