% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_add_variables}
\alias{wq_add_variables}
\title{Adds Variable Column}
\usage{
wq_add_variables(x)
}
\arguments{
\item{x}{data.frame with Code column}
}
\value{
data.frame with Variable and Code columns
}
\description{
Adds a Variable column to the data.frame x based on the Code column.
Unknown variable names are dropped.
}

