% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_codes}
\alias{wq_codes}
\title{Get Water Quality Codes}
\usage{
wq_codes()
}
\description{
Returns a character vector of the water quality codes for which
guidelines are currently defined in the wqbc package.
}
\examples{
wq_codes()
}

