% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wq_index.R
\name{wq_index}
\alias{wq_index}
\title{Water Quality Index}
\usage{
wq_index(x)
}
\arguments{
\item{x}{data.frame with columns Parameter, Value and Threshold}
}
\description{
Calculates water quality index from data.frame with columns Parameter, Value
and Threshold.
}

