% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_jurisdictions}
\alias{wq_jurisdictions}
\title{Get Water Quality Jurisdictions}
\usage{
wq_jurisdictions()
}
\description{
Returns a character vector of the jurisdictions for which
limits are currently defined in the wqbc package.
}
\examples{
wq_jurisdictions()
}

