% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wq_thresholds.R
\name{wq_thresholds}
\alias{wq_thresholds}
\title{Water Quality Thresholds}
\usage{
wq_thresholds(x)
}
\arguments{
\item{x}{data.frame with columns Parameter and Value}
}
\description{
Gets water quality thresholds for Parameters
}

