% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_uses}
\alias{wq_uses}
\title{Get Water Quality Uses}
\usage{
wq_uses()
}
\description{
Returns a character vector of the uses for which
guidelines are currently defined in the wqbc package.
}
\examples{
wq_uses()
}

