% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{wq_variables}
\alias{wq_variables}
\title{Get Water Quality Variables}
\usage{
wq_variables()
}
\description{
Returns a character vector of the water quality variables for which
limits are currently defined in the wqbc package.
}
\examples{
wq_variables()
}

