% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqbc-package.R
\docType{package}
\name{wqbc-package}
\alias{wqbc}
\alias{wqbc-package}
\title{wqbc: Tidy Water Quality Data and Calculate Thresholds for British Columbia}
\description{
Tidies water quality data and calculates water quality thresholds for British Columbia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/wqbc}
  \item Report bugs at \url{https://github.com/bcgov/wqbc/issues}
}

}
\author{
\strong{Maintainer}: Karly Harker \email{karly.harker@gov.bc.ca} [contributor]

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID}) [contractor]
  \item Colin Millar \email{colinpmillar@gmail.com} [contractor]
  \item Andy Teucher \email{andy.teucher@gov.bc.ca}
}

Other contributors:
\itemize{
  \item Sebastian Dalgarno \email{seb@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-3658-4517}{ORCID}) [contributor]
  \item Wendy Wang [contributor, contractor]
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca} [contributor]
  \item Robyn Irvine [contributor, contractor]
  \item Province of British Columbia [copyright holder]
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
}

}
\keyword{internal}
