% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wqbc}
\alias{wqbc}
\alias{wqbc-package}
\title{Water Quality Thresholds and Index Calculation for British Columbia}
\description{
Water Quality Thresholds and Index Calculation for British Columbia
}
\examples{
library(ggplot2)
data(fraser)
plot_map(fraser)

data(ccme)
calc_wqis(ccme)
wqis <- calc_wqis(ccme, by = "Date")
plot_wqis(wqis, x = "Date")
}

