% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-limits.R
\name{calc_limits}
\alias{calc_limits}
\title{Gets Water Quality limits}
\usage{
calc_limits(x, use = "Freshwater Life")
}
\arguments{
\item{x}{data.frame with columns Code and Value}

\item{use}{string of required use}
}
\description{
Gets lower and upper water quality thresholds for British Columbia
}

