% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-limits.R
\name{calc_limits}
\alias{calc_limits}
\title{Calculates Water Quality limits}
\usage{
calc_limits(x, by = NULL, messages = getOption("wqbc.messages", default =
  TRUE), parallel = getOption("wqbc.parallel", default = FALSE))
}
\arguments{
\item{x}{data.frame with columns Code, Value and Units.}

\item{by}{character vector of columns to calculate limits by}

\item{messages}{flag indicating whether to print messages}

\item{parallel}{flag indicating whether to calculate limits by the by argument using the parallel backend provided by foreach
data(fraser)
fraser <- calc_limits(rbind(fraser[1:100,],fraser[1:100,]), message = FALSE)}
}
\description{
Calculates the approved lower and upper water quality thresholds for
British Columbia. If the Date column is not
supplied the data is assumed to have been collected on the same date.
Assumes values are individual readings.
}

