% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-limits.R
\name{calc_limits}
\alias{calc_limits}
\title{Calculates Water Quality limits}
\usage{
calc_limits(x, by = NULL, term = "long", dates = NULL,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The data.frame to perform the calculations on.}

\item{by}{A character vector of the columns to perform the calculations by.}

\item{term}{A string indicating whether to calculate long-term or short-term limits.}

\item{dates}{A date vector indicating the start of each 30 day period.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Calculates the approved upper water quality thresholds for
British Columbia.
}
\examples{
\dontrun{
demo(fraser, ask = FALSE)
}
}

