% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{calc_wqis}
\alias{calc_wqis}
\title{Calculate Water Quality Indices (WQIs)}
\usage{
calc_wqis(x, by = NULL)
}
\arguments{
\item{x}{data.frame with Code, Value, UpperLimit and if defined
LowerLimit columns}

\item{by}{character vector of columns to calculate WQIs by.}
}
\description{
Calculates WQIs for x.
}
\examples{
data(ccme)
calc_wqis(ccme)
calc_wqis(ccme, by = "Date")

require(ggplot2)
wqis <- calc_wqis(ccme, by = "Date")
gp <- ggplot(data = wqis, aes(x = Date, y = WQI))
gp <- gp + geom_line()
gp <- gp + geom_point(aes(size = Tests, color = Category))
gp <- gp + expand_limits(y = 0)
gp <- gp + scale_color_manual(values = get_category_colors())

print(gp)
}

