% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{calc_wqis}
\alias{calc_wqis}
\title{Calculate Water Quality Indices (WQIs)}
\usage{
calc_wqis(x, by = NULL, messages = getOption("wqbc.messages", default =
  TRUE), parallel = getOption("wqbc.parallel", default = FALSE))
}
\arguments{
\item{x}{data.frame with Variable, Value, UpperLimit and if defined
LowerLimit columns}

\item{by}{character vector of columns to calculate WQIs by}

\item{messages}{flag indicating whether to print messages}

\item{parallel}{flag indicating whether to calculate limits by the by argument using the parallel backend provided by foreach}
}
\description{
Calculates WQIs for x.
}
\examples{
data(ccme)
calc_wqis(ccme)
calc_wqis(ccme, by = "Date")
}

