% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{calc_wqis}
\alias{calc_wqis}
\title{Calculate Water Quality Indices (WQIs)}
\usage{
calc_wqis(x, by = NULL)
}
\arguments{
\item{x}{data.frame with Code, Value, UpperLimit and if defined
LowerLimit columns}

\item{by}{character vector of columns to calculate WQIs by.}
}
\description{
Calculates WQIs for x.
}
\examples{
data(ccme)
calc_wqis(ccme)
}

