% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc-wqi.R
\name{categorize_wqi}
\alias{categorize_wqi}
\title{Categorize Water Quality Indices}
\usage{
categorize_wqi(x)
}
\arguments{
\item{x}{A numeric vector of the WQI values to categorize.}
}
\description{
Categorizes numbers between 1 and 100 into
the categories defined in the CCME Water Quality Index 1.0 User's Manual.
}
\examples{
categorize_wqi(seq(1,100,by = 5))
}

