% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-wqis.R
\name{categorize_wqi}
\alias{categorize_wqi}
\title{Categorize Water Quality Indices}
\usage{
categorize_wqi(x)
}
\arguments{
\item{x}{A numeric vector of the WQI values to categorize.}
}
\description{
Categorizes WQI values between 1 and 100 into
the CCME categories.
}
\examples{
categorize_wqi(seq(1,100,by = 5))
}

