% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccme}
\alias{ccme}
\title{Tidy CCME Water Quality Index User’s Manual Example Data}
\format{A data frame with 12 rows and 11 variables:
\describe{
  \item{Code}{parameter code}
  \item{Date}{date of reading}
  \item{Value}{value of reading}
  \item{Guideline}{guideline}
}}
\usage{
ccme
}
\description{
Example tidy dataset from the CCME Water Quality Index 1.0 User’s Manual
}
\keyword{datasets}

