% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ccme}
\alias{ccme}
\title{CCME Water Quality Index User's Manual Example Data}
\format{A data frame with 103 rows and 6 variables:
\describe{
  \item{Code}{parameter code}
  \item{Date}{date of reading}
  \item{Value}{value of reading}
  \item{DetectionLimit}{detection limit of method}
  \item{LowerLimit}{minimum permitted value}
  \item{UpperLimit}{maximum permitted value}
}}
\usage{
ccme
}
\description{
A tidy data.frame of the CCME Water Quality Index 1.0 User's Manual
example dataset.
}
\keyword{datasets}

