% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean-wqdata.R
\name{clean_wqdata}
\alias{clean_wqdata}
\title{Clean Water Quality Data}
\usage{
clean_wqdata(x, by = NULL, max_cv = 1.29,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The data.frame to clean.}

\item{by}{A character vector of the columns in x to perform the cleaning by.}

\item{max_cv}{A number indicating the maximum permitted coefficient
of variation for replicates.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Cleans water quality data. After standardization using \code{\link{standardize_wqdata}}
replicates (two or more readings for the same variable on the same date) are averaged
using the \code{mean} function.
Readings for the same variable on the same date but at different levels of the
columns specified in by are not considered replicates. The \code{clean_wqdata}
function is automatically called by \code{\link{calc_limits}} prior
to calculating limits.
}
\details{
If there are three or more replicates with a coefficient of variation (CV) in
exceedance of \code{max_cv} then the replicates with the highest absolute deviation
is dropped until the CV is less than or equal to \code{max_cv}
or only two values remain. The default value max_cv value of 1.29
is exceeded by two zero and one positive value (CV = 1.73)
or by two identical positive values and a third value an order
or magnitude greater (CV = 1.30). It is not exceed by one zero
and two identical positive values (CV = 0.87).
}
\examples{
clean_wqdata(wqbc::dummy, messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}

