% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-wqdata.R
\name{clean_wqdata}
\alias{clean_wqdata}
\title{Clean Water Quality Data}
\usage{
clean_wqdata(
  x,
  by = NULL,
  max_cv = Inf,
  sds = 10,
  ignore_undetected = TRUE,
  large_only = TRUE,
  delete_outliers = FALSE,
  remove_blanks = FALSE,
  messages = getOption("wqbc.messages", default = TRUE),
  FUN = mean
)
}
\arguments{
\item{x}{The data.frame to clean.}

\item{by}{A character vector of the columns in x to perform the cleaning by.
If you have multiple stations specify the column name that contains the station IDs.}

\item{max_cv}{A number indicating the maximum permitted coefficient
of variation for replicates.}

\item{sds}{The number of standard deviations above which a value is considered an outlier.}

\item{ignore_undetected}{A flag indicating whether to ignore undetected values when calculating the average deviation and identifying outliers.}

\item{large_only}{A flag indicating whether only large values which exceed the sds should be identified as outliers.}

\item{delete_outliers}{A flag indicating whether to delete outliers or merely flag them.}

\item{remove_blanks}{Should blanks be removed? Blanks are assumed to be denoted by
a value of \code{"Blank..."} in the \code{SAMPLE_CLASS} column. Default \code{FALSE}}

\item{messages}{A flag indicating whether to print messages.}

\item{FUN}{The function to use for summaries, e.g. \code{median}, \code{mean}, or \code{max}. Default \code{mean}}
}
\description{
Cleans water quality data. After standardization using \code{\link{standardize_wqdata}}
replicates (two or more readings for the same variable on the same date) are averaged
using the \code{mean} function.
Readings for the same variable on the same date but at different levels of the
columns specified in by are not considered replicates. The \code{clean_wqdata}
function is automatically called by \code{\link{calc_limits}} prior
to calculating limits.
}
\details{
If there are three or more replicates with a coefficient of variation (CV) in
exceedance of \code{max_cv} then the replicates with the highest absolute deviation
is dropped until the CV is less than or equal to \code{max_cv}
or only two values remain. By default all values are averaged.

A max_cv value of 1.29
is exceeded by two zero and one positive value (CV = 1.73)
or by two identical positive values and a third value an order
or magnitude greater (CV = 1.30). It is not exceed by one zero
and two identical positive values (CV = 0.87).
}
\examples{
clean_wqdata(wqbc::dummy, messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}
