% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\title{Water Quality Parameter Codes and Units for British Columbia}
\format{A data frame with 4 variables:
\describe{
  \item{Variable}{The name of the variable.}
  \item{Code}{The EMS code in expanded form.}
  \item{Units}{The standard units for the variable.}
  \item{Average}{R function to calculate "average" value.}
  \item{EC_Code}{The Variable Code in the Environment Canada data.}
}}
\usage{
codes
}
\description{
The standard variables and codes recognised by the wqbc package with their
standard units and the R function to use when averaging multiple samples.
}
\seealso{
\code{\link{calc_limits}}
}
\keyword{datasets}
