% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/codes.R
\name{compress_ems_codes}
\alias{compress_ems_codes}
\title{Compress EMS Codes}
\usage{
compress_ems_codes(x)
}
\arguments{
\item{x}{A character vector of codes to compress.}
}
\description{
Compresses EMS codes by removing EMS_ from
start and replacing all '_' with '-'.
}
\examples{
compress_ems_codes(c("EMS_0014", "EMS_KR-P", "0-15"))
}

