% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{convert_values}
\alias{convert_values}
\title{Convert values to different units}
\usage{
convert_values(x, from, to, messages)
}
\arguments{
\item{x}{a numeric vector of values to convert}

\item{from}{units to convert from}

\item{to}{units to convert to}

\item{messages}{should messages be printed when}
}
\value{
a numeric vector of the converted values
}
\description{
Convert values to different units
}
\details{
Currently supported units for \code{from} and \code{to} are:
c("ng/L", "ug/L", "mg/L", "g/L", "kg/L", "pH", "degC", "C", "CFU/dL", "MPN/dL", "CFU/100mL", "MPN/100mL", "CFU/g", "MPN/g", "CFU/mL", "MPN/mL", "Col.unit", "Rel", "NTU", "m", "uS/cm")
}
\examples{

convert_values(1, "ug/L", "mg/L", messages = FALSE)

df <- data.frame(
  value = c(1.256, 5400000, 12300, .00098),
  units = c("mg/L", "ng/L", "ug/L", "g/L"),
  stringsAsFactors = FALSE
)
df

df$units_mg_L <- convert_values(df$value, from = df$units, to = "mg/L", messages = FALSE)
df
}
