% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dummy}
\alias{dummy}
\title{Dummy Water Quality Data}
\format{A data frame with 5 columns:
\describe{
  \item{ID}{A row identifier.}
  \item{Date}{The date of the reading.}
  \item{Variable}{The name of the variable.}
  \item{Value}{The value of the reading.}
  \item{Units}{The value's units.}
}}
\usage{
dummy
}
\description{
A dummy data set to illustrate various data cleaning functions.
}
\examples{
demo(dummy, ask = FALSE)
}
\keyword{datasets}

