% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fraser}
\alias{fraser}
\title{Fraser River Basin Long-term Water Quality Monitoring 1979-Present}
\format{A data frame with 9 variables:
\describe{
  \item{SiteID}{unique water quality station number}
  \item{Date}{date of water quality sample}
  \item{Variable}{variable name}
  \item{Value}{measured value}
  \item{Units}{reading units}
  \item{DetectionLimit}{minimum value of method}
  \item{Site}{full station name}
  \item{Lat}{latitude in decimal degrees}
  \item{Long}{longitude in decimal degrees}
}}
\source{
\url{http://open.canada.ca/data/en/dataset/9ec91c92-22f8-4520-8b2c-0f1cce663e18}
}
\usage{
fraser
}
\description{
Surface freshwater quality monitoring in the Fraser River Basin
is carried out under the Canada-British Columbia Water Quality
Monitoring Agreement. Monitoring is conducted to assess water
quality status and long-term trends, detect emerging issues,
establish water quality guidelines and track the effectiveness
of remedial measures and regulatory decisions.
}
\details{
The original dataset has been filtered to remove values
for variables without currently defined limits. In addition,
variables are referenced
by code, unimportant columns have been dropped and the remaining
columns renamed.
}
\keyword{datasets}

