% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{get_codes}
\alias{get_codes}
\title{Get Codes}
\usage{
get_codes(variables = NULL, add_na = TRUE)
}
\arguments{
\item{variables}{An optional character vector of variables to get codes for.}

\item{add_na}{A flag indicating whether to replace variables without codes
with missing values.}
}
\description{
Gets the recognised water quality codes.
}
\examples{
get_codes()
}

