% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{get_codes}
\alias{get_codes}
\title{Get Codes}
\usage{
get_codes(variables = NULL, messages = getOption("wqbc.messages", default =
  TRUE))
}
\arguments{
\item{variables}{An optional character vector of variables to get codes for.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Gets the recognised water quality codes.
}
\examples{
get_codes()
get_codes(c(get_variables()[1:3], "Kryptonite"))
}

