% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_codes}
\alias{get_codes}
\title{Get Water Quality Codes}
\usage{
get_codes(variables = NULL, add_na = TRUE)
}
\arguments{
\item{variables}{optional character vector of variables to get codes for}

\item{add_na}{flag indicating whether to replace variables without codes
with NAs}
}
\description{
Returns a character vector of the water quality codes for which
limits are currently defined in the wqbc package.
}
\examples{
get_codes()
variables <- c("Silver", "Kryptonite", "Arsenic", NA, "pH", "Total Phosphorus")
get_codes(variables)
get_codes(variables, add_na = FALSE)
}

