% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{get_variables}
\alias{get_variables}
\title{Get Variables}
\usage{
get_variables(codes = NULL)
}
\arguments{
\item{codes}{An optional character vector of the codes to get the variables for.}
}
\value{
A character vector of the water quality variables.
}
\description{
Gets recognised water quality variables.
}
\examples{
get_variables()
get_variables(get_codes())
get_variables(c(get_codes()[1], "EMS_WTF_"))
}

