% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_variables}
\alias{get_variables}
\title{Get Water Quality Variables}
\usage{
get_variables(codes = NULL)
}
\arguments{
\item{codes}{optional character vector of codes to get variables for}
}
\description{
Returns a character vector of the water quality variables for which
limits are currently defined in the wqbc package.
}
\examples{
get_variables()
get_variables(c("Ag", "KR", "As", NA, "pH", "TP"))
}

