% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{get_variables}
\alias{get_variables}
\title{Get Water Quality Variables}
\usage{
get_variables(codes = NULL)
}
\arguments{
\item{codes}{optional character vector of codes to get variables for}
}
\description{
Returns a character vector of the water quality variables
recognised by the wqbc package.
}
\examples{
get_variables()
}

