% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify-outliers.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify Outliers In Water Quality Data}
\usage{
identify_outliers(data, by = NULL, sds = 6, ignore_undetected = TRUE,
  large_only = TRUE, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{A data.frame with a numeric Value column to identify the outliers for.}

\item{by}{A character vector of the columns in x to perform the outlier detection by.}

\item{sds}{The number of standard deviations above which a value is considered an outlier.}

\item{ignore_undetected}{A flag indicating whether to ignore undetected values when calculating the average deviation and identifying outliers.}

\item{large_only}{A flag indicating wether only large values which exceed the sds should be identified as outliers.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
The data with an additional logical column Outlier which indicates which values were identified as outliers.
}
\description{
Identifies outliers in water quality data based on the standard deviation.
}
\examples{
data <- standardize_wqdata(wqbc::dummy)
identify_outliers(data, by = "Variable", messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}
