% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limits}
\alias{limits}
\title{Water Quality Limits for British Columbia and Canada}
\format{A data frame with 13 variables:
\describe{
  \item{Variable}{name of water quality parameter}
  \item{Term}{Period for which limit applies. Either Short which
  indicates any individual measurement or Long which indicates
  average of at least 5 values within a 30 day period.}
  \item{Condition}{R logical expression to test required condition}
  \item{UpperLimit}{R expression defining upper limit}
  \item{Units}{units for limit}
}}
\usage{
limits
}
\description{
Water Quality Limits for British Columbia and Canada
}
\keyword{datasets}

