% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limits}
\alias{limits}
\title{Water Quality Limits for British Columbia and Canada}
\format{A data frame with 13 variables:
\describe{
  \item{Code}{short-hand code}
  \item{Variable}{name of water quality parameter}
  \item{Jurisdiction}{regulatory jurisdiction in Canada (permitted values:
  BC or CA)}
  \item{Use}{intended use (permitted values: Drinking, Freshwater Life,
   Marine Life, Wildlife, Livestock, Irrigation or Recreation)}
  \item{SubUse}{additional information on use}
  \item{Samples}{minimum number of samples required}
  \item{Period}{period within which number of samples must be collected (permitted values: Day or Month)}
  \item{Average}{R function(s) to calculate "average" value for multiple samples}
  \item{Condition}{R logical expression to test required condition}
  \item{LowerLimit}{R expression defining lower limit}
  \item{UpperLimit}{R expression defining upper limit}
  \item{Units}{units for guideline (permitted values: mg/L, ug/L, /dL, m, NTU or pH)}
}}
\usage{
limits
}
\description{
Both the Canadian federal government and the
province of British Columbia set limits for a range of
water quality parameters. The data were taken from a
range of federal and provincial websites. They represent
long-term limits for samples taken from the water column.
Limits which depend on conditions from other periods
or locations are not included. For example some of the
limits for turbidity are defined with respect to background
levels and/or an upstream site.
Where some interpretation of the information provided was required
this is noted in the comments.
}
\keyword{datasets}

