% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limits}
\alias{limits}
\title{Water Quality Limits for British Columbia and Canada}
\format{A data frame with 13 variables:
\describe{
  \item{Code}{short-hand code}
  \item{Variable}{name of water quality parameter}
  \item{Average}{R function(s) to calculate "average" value for multiple samples}
  \item{Condition}{R logical expression to test required condition}
  \item{LowerLimit}{R expression defining lower limit}
  \item{UpperLimit}{R expression defining upper limit}
  \item{Units}{units for limit}
  \item{Form}{variable form}
}}
\usage{
limits
}
\description{
Water Quality Limits for British Columbia and Canada
}
\keyword{datasets}

