% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limits}
\alias{limits}
\title{Water Quality Limits for British Columbia}
\format{
A data frame with 6 variables:
\describe{
\item{Variable}{The name of the variable.}
\item{Use}{The name of the Use.}
\item{Term}{The term of the limit i.e. "Short" versus "Long".}
\item{Condition}{A logical R expression to test for the required condition.}
\item{UpperLimit}{The upper limit or an R expression defining the upper limit.}
\item{Units}{The units of the upper limit.}
}
}
\usage{
limits
}
\description{
The short and long term water quality limits for British Columbia recognised
by the wqbc package.
}
\seealso{
\code{\link{calc_limits}}
}
\keyword{datasets}
