% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_variables}
\alias{lookup_variables}
\title{Lookup Variables}
\usage{
lookup_variables(codes = NULL, messages = getOption("wqbc.messages", default
  = TRUE))
}
\arguments{
\item{codes}{An optional character vector of the codes to get the variables for.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Looks up recognised water quality variables.
Returns character vector of the water quality variables.
}
\examples{
lookup_variables()
lookup_variables(lookup_codes())
lookup_variables(c(lookup_codes()[1:3], "KRYP"))
}

