% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot Map}
\usage{
plot_map(data, x = "Long", y = "Lat", size = 3, shape = 21, fill = 10,
  theme = theme_map(), drop = TRUE, input_proj = NULL)
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{y}{string of column in data to plot on y axis}

\item{size}{number of point size or string of column in data to plot size of points}

\item{shape}{integer of point shape (permitted values are 21 to 25) or string of column in data to plot shape of points}

\item{fill}{integer of point fill or string of column in data to plot fill of points}

\item{theme}{ggplot theme}

\item{drop}{flag indicating whether to drop duplicated rows to
avoid overplotting}

\item{input_proj}{a valid proj4string. Defaults to longlat/NAD83
(\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"})}
}
\description{
Creates ggplot2 object with map polygon,
coord_fixed and theme_minimal.
}
\examples{
library(ggplot2)
library(sp)
library(rgdal)

data(fraser)
plot_map(fraser)

\dontrun{
 demo(fraser)
}
}

