% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot Map}
\usage{
plot_map(data, x = "Long", y = "Lat", size = 3, shape = 21, fill = 10,
  keep = NULL, input_proj = NULL)
}
\arguments{
\item{data}{A data.frame with spatial information to map.}

\item{x}{A string of the column in data to plot on the x axis.}

\item{y}{A string of the column in data to plot on the y axis.}

\item{size}{A number of the point size or a string of the column in data
to represent by the size of points.}

\item{shape}{An integer of the point shape (permitted values are 21 to 25)
or a string of the column in data to represent by the shape of points.}

\item{fill}{An integer of the point fill colour or a string of the column in data to represent
by the fill colour of points.}

\item{keep}{An optional character vector indicating which columns
in addition to x and y to keep before dropping duplicated rows to
avoid overplotting.}

\item{input_proj}{An optional valid proj4string. Defaults to
(\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"}).}
}
\description{
Creates a ggplot2 object with a polygon of British Columbia.
If any columns are required
for additional layers in the plot or facetting then
they should be specified in the keep argument.
}
\examples{
library(ggplot2)
library(sp)
library(rgdal)

data(fraser)
plot_map(fraser)

\dontrun{
 demo(fraser)
}
}
\seealso{
\code{\link{plot_map_wqis}}
}
