% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map}
\alias{plot_map}
\title{Plot Map}
\usage{
plot_map(data, x = "Longitude", y = "Latitude", input_proj = NULL)
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{y}{string of column in data to plot on y axis}

\item{input_proj}{a valid proj4string. Defaults to longlat/NAD83 (\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"})}
}
\value{
ggplot2 object
}
\description{
Creates ggplot2 object with map polygon,
coord_fixed and theme_minimal.
}
\examples{
library(ggplot2)
data(fraser)
plot_map(fraser)
}

