% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map_wqis}
\alias{plot_map_wqis}
\title{Plot Map of Water Quality Index Categories.}
\usage{
plot_map_wqis(data, x = "Long", y = "Lat", size = 3, shape = 21,
  keep = NULL, input_proj = NULL)
}
\arguments{
\item{data}{A data.frame of WQI values to plot.}

\item{x}{A string of the column in data to plot on the x axis.}

\item{y}{A string of the column in data to plot on the y axis.}

\item{size}{A number of the point size or string of the column in data
to represent by the size of points.}

\item{shape}{An integer of the point shape (permitted values are 21 to 25)
or string of the column in data to represent by the shape of points.}

\item{keep}{An optional character vector indicating which columns to keep
in addition to x and y to keep before dropping duplicated rows to
avoid overplotting.}

\item{input_proj}{An optional valid proj4string. Defaults to
(\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"}).}
}
\description{
Creates a ggplot2 object with a polygon of British Columbia with
the Water Quality Index categories indicated by the fill colour of points.
}
\examples{
\dontrun{
 demo(fraser)
}
}
\seealso{
\code{\link{plot_wqis}} and \code{\link{plot_map}}
}

