% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_map_wqis}
\alias{plot_map_wqis}
\title{Plot Map of Water Quality Indices}
\usage{
plot_map_wqis(data, x = "Long", y = "Lat", size = 3, shape = 21,
  theme = theme_map(), keep = NULL, palette = "gradient",
  input_proj = NULL)
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{y}{string of column in data to plot on y axis}

\item{size}{number of point size or string of column in data
to plot size of points}

\item{shape}{integer of point shape (permitted values are 21 to 25)
or string of column in data to plot shape of points}

\item{theme}{ggplot theme}

\item{keep}{NULL or character vector indicating which columns to keep
in addition to x and y before avoiding overplotting by dropping duplicated rows.}

\item{palette}{string of \code{\link{get_category_colours}} palette to use.}

\item{input_proj}{a valid proj4string. Defaults to longlat/NAD83
(\code{"+proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +no_defs"})}
}
\description{
Creates ggplot2 object with map polygon,
coord_fixed and theme_minimal and fill based
on WQI categories.
}
\examples{
\dontrun{
 demo(fraser)
}
}
\seealso{
\code{\link{plot_map}}
}

