% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot Time Series}
\usage{
plot_timeseries(data, x = "Date", y = "Value", dl = "DetectionLimit",
  xlab = x, ylab = y, title = NULL, color = "black", y0 = TRUE)
}
\arguments{
\item{data}{A data frame of the data to plot.}

\item{x}{A string of the column to plot on the x-axis.}

\item{y}{A string of the column to plot on the y-axis.}

\item{dl}{A string of the column that provides the detection limit(s).}

\item{xlab}{A string of the x-axis label.}

\item{ylab}{A string of the y-axis label.}

\item{title}{A string of the plot title.}

\item{color}{A string specifying the color for the points or if not a color the name of the column to color the points by.}

\item{y0}{A flag indicating whether to expand the y-axis limits to include 0.}
}
\description{
The plot includes the detection limits.
To exclude detection limits. Set \code{dl = NULL}.
}
\examples{
plot_timeseries(ccme[ccme$Variable == "As",])
}

