% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot Time Series Data}
\usage{
plot_timeseries(
  data,
  by = NULL,
  y0 = TRUE,
  size = 1,
  messages = getOption("wqbc.messages", default = TRUE)
)
}
\arguments{
\item{data}{A data frame of the data to plot.}

\item{by}{A character vector of the columns to plot the time series by.}

\item{y0}{A flag indicating whether to expand the y-axis limits to include 0.}

\item{size}{A number of the point size.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
If \code{by = NULL} plot_timeseries returns a ggplot object.
Otherwise it returns a list of ggplot objects.
}
\examples{
plot_timeseries(ccme[ccme$Variable == "As",])
plot_timeseries(ccme, by = "Variable")
}
