% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wqis}
\alias{plot_wqis}
\title{Plot Water Quality Indices}
\usage{
plot_wqis(data, x = "Tests", size = NULL, shape = NULL)
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{size}{string of column in data to plot size of points}

\item{shape}{string of column in data to plot shape of points}
}
\value{
ggplot2 object
}
\description{
Creates ggplot2 object with map polygon,
limits expanded to include 0 and 100 and color scale
with values get_category_colors()
}
\examples{
library(ggplot2)

data(ccme)
plot_wqis(calc_wqis(ccme))

wqis <- calc_wqis(ccme, by = "Date")
plot_wqis(wqis)
plot_wqis(wqis, x = "Date")

library(lubridate)

wqis$Year <- year(wqis$Date)
wqis$Dayte <- wqis$Date
year(wqis$Dayte) <- 2000
plot_wqis(wqis, x = "Dayte", size = "Tests") +
facet_wrap(~Year) + xlab("Day of the Year") + theme_bw()
}

