% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wqis}
\alias{plot_wqis}
\title{Plot Water Quality Indices}
\usage{
plot_wqis(data, x = "Tests")
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}
}
\value{
ggplot2 object
}
\description{
Creates ggplot2 object with map polygon,
limits expanded to include 0 and 100 and color scale
with values get_category_colors()
}
\examples{
library(ggplot2)
data(ccme)
plot_wqis(calc_wqis(ccme))
plot_wqis(calc_wqis(ccme, by = "Date"))
plot_wqis(calc_wqis(ccme, by = "Date"), x = "Date")
}

