% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_wqis}
\alias{plot_wqis}
\title{Plot Water Quality Indices}
\usage{
plot_wqis(data, x = "Tests", size = 3, shape = 21, palette = "default",
  theme = theme_wqis())
}
\arguments{
\item{data}{data.frame to plot}

\item{x}{string of column in data to plot on x axis}

\item{size}{number of size or string of column in data to plot size of points}

\item{shape}{integer of shape (permitted values are 21 to 25) or string of column in data to plot shape of points}

\item{palette}{string indicating palette to use. Possible values are
"default" and "blue"}

\item{theme}{ggplot theme}
}
\value{
ggplot2 object
}
\description{
Creates ggplot2 object with map polygon,
limits expanded to include 0 and 100 and colour scale
with values get_category_colours()
}
\examples{
library(ggplot2)

data(ccme)
plot_wqis(calc_wqis(ccme))

wqis <- calc_wqis(ccme, by = "Date")
plot_wqis(wqis)
plot_wqis(wqis, x = "Date")

library(lubridate)

wqis <- calc_wqis(ccme, by = "Date")
wqis$Year <- year(wqis$Date)
wqis$Dayte <- wqis$Date
year(wqis$Dayte) <- 2000
plot_wqis(wqis, x = "Dayte", size = "Tests") +
facet_wrap(~Year) + xlab("Day of the Year") + theme_bw()

test <- data.frame(WQI = seq(0, 100, by = 5))
test$Category = categorize_wqi(test$WQI)
for (palette in c("default", "blue")) {
 print(plot_wqis(test, x = "WQI", palette = palette) + xlab(palette))
}
}

