% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize-wqdata.R
\name{standardize_wqdata}
\alias{standardize_wqdata}
\alias{standardise_wqdata}
\title{Standardize Water Quality Data}
\usage{
standardize_wqdata(x, strict = TRUE,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{A data.frame to standardize.}

\item{strict}{A flag that is passed to substitute_variables indicating
whether to require all words in a recognised variable name to be
present in x (strict = TRUE) or only the first one (strict = FALSE)}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Standardizes a water quality data set using \code{\link{substitute_variables}}
and \code{\link{substitute_units}} so that all remaining values have the
recognised codes and variables in \code{\link{codes}} and the standard
units. If column Code is present then a Variable column is created using
\code{\link{lookup_variables}}. The \code{standardize_wqdata} function
is called by \code{clean_wqdata} prior to cleaning.
}
\examples{
standardize_wqdata(wqbc::dummy, messages = TRUE)
}
\seealso{
\code{\link{clean_wqdata}}
}
