% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_units}
\alias{substitute_units}
\title{Substitute Units}
\usage{
substitute_units(x, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The character vector of units to substitute.}

\item{messages}{A flag indicating whether to print messages.}
}
\description{
Substitutes provided unit names for recognised units.
Before matching all spaces and "units" or "UNITS" are removed.
The case is not important. Where
there are no matches missing values are returned.
}
\examples{
substitute_units(c("mg/L", "MG/L", "mg /L ", "Kg/l", "gkl"), messages = TRUE)
}
\seealso{
\code{\link{substitute_variables}}
}
