% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/units.R
\name{substitute_units}
\alias{substitute_units}
\title{Substitute Units}
\usage{
substitute_units(x, messages = TRUE)
}
\arguments{
\item{x}{character vector of units to substitute}

\item{messages}{flag indicating whether to provide messages}
}
\value{
character vector of substituted units where
match or NA
}
\description{
Where possible substitute units with
possible values
}
\examples{
substitute_units(c("mg/L", "MG/L", "mg /L ", "Kg/l", "kg.l"), FALSE)
}

