% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/units.R
\name{substitute_units}
\alias{substitute_units}
\title{Substitute Units}
\usage{
substitute_units(x, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{x}{The character vector of units to substitute.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A character vector of the substituted or original units.
}
\description{
Where possible substitute units with
recognised values.
}
\examples{
substitute_units(c("mg/L", "MG/L", "mg /L ", "Kg/l", "gkl", "CFU/100ML"))
substitute_units(c("MG/L", "MG/L", "MG/L"))
substitute_units("gkl")
substitute_units(c(NA, "mg/L"))
}

