% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_variables}
\alias{substitute_variables}
\title{Substitute Variable Names}
\usage{
substitute_variables(x, strict = TRUE, messages = TRUE)
}
\arguments{
\item{x}{character vector of variable names to substitute}

\item{strict}{flag indicating whether to require all words
in a variable name to be present in (strict = TRUE) or only
the first word (strict = FALSE). Note when strict = FALSE
ambiguous variables such as "Polychlorinated Biphenyl Total"
and "Polychlorinated Biphenyl 105" are dropped.}

\item{messages}{flag indicating whether to print messages}
}
\value{
character vector of substituted or original names
}
\description{
Where possible substitute variable names
with recognised variable names
}
\examples{
substitute_variables(c("ALUMINIUM SOMETHING", "FLUORIDE DISSOLVED",
"FLUORIDE", "NITROGEN DISSOLVED NITRATE", "PHOSPHORUS - TOTAL",
"KRYPTONITE", "OXYGEN", "OXYGEN DISSOLVED"))
}

