% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/substitute.R
\name{substitute_variables}
\alias{substitute_variables}
\title{Substitute Variable Names}
\usage{
substitute_variables(x, messages = TRUE)
}
\arguments{
\item{x}{character vector of variable names to substitute}

\item{messages}{flag indicating whether to print messages}
}
\value{
character vector of substituted or original names
}
\description{
Where possible substitute variable names
with recognised variable names
}
\examples{
substitute_variables(c("ALUMINIUM SOMETHING", "FLUORIDE DISSOLVED",
"FLUORIDE", "NITROGEN DISSOLVED NITRATE", "PHOSPHORUS - TOTAL",
"KRYPTONITE", "OXYGEN", "OXYGEN DISSOLVED"))
}

