% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-trends.R
\name{summarise_for_trends}
\alias{summarise_for_trends}
\title{Summarise data by year and month}
\usage{
summarise_for_trends(data, breaks = NULL, FUN = "median",
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{breaks}{A numeric vector used to create groups of consecutive months, if NULL the full
year is used.}

\item{FUN}{The function to use for yearly summaries, e.g. median, mean, or max.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A tibble data.frame with rows for each Station, Variable, Year and month grouping.
}
\description{
Compute annual summaries of water quality observations.
}
\details{
The data must contain the columns Station, Date, Variable, Value, and Units.
}
\examples{
 # select one station
 data(yuepilon)
 data <- yuepilon[yuepilon$Station == "02EA005", ]
 # estimate trend (using simple sen slope)
 trend <- test_trends(data, messages = TRUE)
 # get the data used in the test
 datasum <- summarise_for_trends(data)
 plot(datasum$Year, datasum$Value,
      main = paste("p-value =", round(trend$significance, 3)),
      ylab = "Value", xlab = "Year", las = 1)
}
