% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-trends.R
\name{test_trends}
\alias{test_trends}
\title{Seasonal Mann-Kendall Trend Teste}
\usage{
test_trends(data, messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A tibble data.frame with The data with an additional logical column Outlier which indicates which values were identified as outliers.
}
\description{
Analyses time series using the Seasonal Mann-Kendall test.
}
\details{
The data must contain the columns Station_Name, Date, Variable, Value, and Units.
}
\examples{
data <- standardize_wqdata(wqbc::dummy)
identify_outliers(data, by = "Variable", messages = TRUE)
}
\seealso{
\code{\link{calc_limits}} and \code{\link{standardize_wqdata}}
}

