% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-trends.R
\name{test_trends}
\alias{test_trends}
\title{Thiel-Sen Trend Teste}
\usage{
test_trends(data, breaks = NULL, FUN = "median", level = 0.95,
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{breaks}{A numeric vector used to create groups of consecutive months, if NULL the full
year is used.}

\item{FUN}{The function to use for yearly summaries, e.g. median, mean, or max.}

\item{level}{The confidence level to compute a confidence interval for the Thiel-Sen slope.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A tibble data.frame with rows for each Station Variable, and month grouping, and
        additional columns for the slope estinate and confidence intervals.
}
\description{
Analyses time series using the Thiel-Sen estimate of slope.
}
\details{
The data must contain the columns Station, Date, Variable, Value, and Units.
}
\examples{
#data <- standardize_wqdata(wqbc::dummy)
#test_trends(data, breaks = 6, messages = TRUE)
}

