% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-trends.R
\name{test_trends}
\alias{test_trends}
\title{Thiel-Sen Trend Test}
\usage{
test_trends(data, breaks = NULL, FUN = "median",
  messages = getOption("wqbc.messages", default = TRUE))
}
\arguments{
\item{data}{The data.frame to analyse.}

\item{breaks}{A numeric vector used to create groups of consecutive months, if NULL the full
year is used.}

\item{FUN}{The function to use for yearly summaries, e.g. median, mean, or max.}

\item{messages}{A flag indicating whether to print messages.}
}
\value{
A tibble data.frame with rows for each Station, Variable, and month grouping, and
additional columns for the sen slope estinate, 95\% confidence intervals on the slope, intercept and significance at the 5\% level for the sen slope.
}
\description{
Analyses time series using the Thiel-Sen estimate of slope. It requires at least 6 years of data.
}
\details{
The data must contain the columns Station, Date, Variable, Value, and Units.
}
\examples{
 data <- wqbc::yuepilon
 trend <- test_trends(data, breaks = 6, messages = TRUE)
\dontrun{
  demo(test_trends)
}
}
\seealso{
\code{\link[zyp]{zyp.sen}}
}
