% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-data.R
\name{tidy_ec_data}
\alias{tidy_ec_data}
\title{Tidy Environment Canada Data}
\usage{
tidy_ec_data(x, cols = character(0), mdl_action = c("zero", "mdl",
  "half", "na", "none"))
}
\arguments{
\item{x}{The data to tidy.}

\item{cols}{additional columns from the EMS data to retain specified as a
character vector of column names that exist in the data.
The dafault columns retained are:
\itemize{
\item "SITE_NO"
\item "DATE_TIME_HEURE" (Renamed to "DateTime")
\item "VARIABLE" (Renamed to "Variable")
\item "VMV_CODE" (Renamed to "Code")
\item "VALUE_VALEUR" (Renamed to "Value")
\item "UNIT_UNITE" (Renamed to "Units")
\item "DSL_LDE" (Renamed to "DetectionLimit")
\item "FLAG_MARQUEUR" (Renamed to "ResultLetter")
}}

\item{mdl_action}{What to do with values below the detection limit. Options
are \code{"zero"} (set the value to \code{0}; the default), #' \code{"half"} (set the value
to half the MDL), \code{"mdl"} (set the value to equal to the MDL), or \code{"na"} (set
the value to \code{NA}). Can also be set to \code{"none"} to leave as is.}
}
\value{
A tibble of the tidied rems data.
}
\description{
Tidies water quality data downloaded from Environment Canada website. It
is recommended to obtain the data via \code{\link[canwqdata:wq_site_data]{canwqdata::wq_site_data()}} or
\code{\link[canwqdata:wq_basin_data]{canwqdata::wq_basin_data()}}
It retains and renames required columns and sets the timezone to PST.
}
