% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wq_index}
\alias{wq_index}
\title{Water Quality Index}
\usage{
wq_index(x)
}
\arguments{
\item{x}{data.frame with columns Parameter, Value and Threshold}
}
\description{
Calculates water quality index from data.frame with columns Parameter, Value
and Threshold.
}

