% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wqbc-package.R
\docType{package}
\name{wqbc-package}
\alias{wqbc}
\alias{wqbc-package}
\title{wqbc: Tidy Water Quality Data and Calculate Thresholds for British Columbia}
\description{
Tidies water quality data and calculates water quality thresholds
for British Columbia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcgov/wqbc}
  \item Report bugs at \url{https://github.com/bcgov/wqbc/issues}
}

}
\author{
\strong{Maintainer}: Andy Teucher \email{andy.teucher@gov.bc.ca}

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (Primary author) [contractor]
  \item Colin Millar \email{colinpmillar@gmail.com} [contractor]
}

Other contributors:
\itemize{
  \item Wendy Wang [contributor, contractor]
  \item Stephanie Hazlitt \email{stephanie.hazlitt@gov.bc.ca} [contributor]
  \item Robyn Irvine [contributor, contractor]
  \item Province of British Columbia [copyright holder]
}

}
\keyword{internal}
