% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wqbc}
\alias{wqbc}
\alias{wqbc-package}
\title{Water Quality Thresholds and Index Calculation for British Columbia}
\description{
Water Quality Thresholds and Index Calculation for British Columbia
}
\examples{
data(ccme)
calc_wqis(ccme)
calc_wqis(ccme, by = "Date")
}
\seealso{
\code{\link{plot_map}} and \code{\link{plot_wqis}}
}

