% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{wqbc}
\alias{wqbc}
\alias{wqbc-package}
\title{Water Quality Thresholds and Index Calculation for British Columbia}
\description{
The wqbc package calculates water quality thresholds and indices for
freshwater life in British Columbia. To the best of the
package creator's knowledge the limits correspond to the
approved Provincial thresholds on the date of
release but wqbc comes with ABSOLUTELY NO WARRANTY. For further
information on the use of the wqbc package please type
\code{vignette("wqbc")} at the R console.
}
\examples{
\dontrun{
 demo(ccme)
 demo(dummy)
 demo(fraser)
}
}
\seealso{
\code{\link{lookup_limits}}, \code{\link{calc_wqi}}
and \code{\link{plot_wqis}}
}

